+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.38.4
 
 - **REFACTOR**: Remove required parameters from `ShadThemeData`, use default values instead.
+- **FEAT**: Add `setInternalFieldError` to `ShadForm` to set a forced error text for a form field.
 
 ## 0.38.3